fix make responsive design for all screen and strong draw and win logic
