+++ src/constants/pieceIcons.ts
@@ -1,0 +1,29 @@
+import bk from "@/assets/bK.svg";
+import bq from "@/assets/bQ.svg";
+import br from "@/assets/bR.svg";
+import bb from "@/assets/bB.svg";
+import bn from "@/assets/bN.svg";
+import bp from "@/assets/bP.svg";
+import wk from "@/assets/wK.svg";
+import wq from "@/assets/wQ.svg";
+import wr from "@/assets/wR.svg";
+import wb from "@/assets/wB.svg";
+import wn from "@/assets/wN.svg";
+import wp from "@/assets/wP.svg";
+
+const pieceIcons: Record<string, string> = {
+  K: wk,
+  Q: wq,
+  R: wr,
+  B: wb,
+  N: wn,
+  P: wp,
+  k: bk,
+  q: bq,
+  r: br,
+  b: bb,
+  n: bn,
+  p: bp,
+};
+
+export default pieceIcons;
