+++ src/pages/chessgame/ChessGame.css
@@ -4,9 +4,9 @@
   flex-direction: column;
   align-items: center;
   padding: 2rem;
-  background: transparent; 
+  background: transparent;
   min-height: 100vh;
-  z-index: 1; /* ensure content stays above background */
+  z-index: 1;
 }
 
 /* Header */
@@ -16,7 +16,7 @@
 }
 
 .title {
-  font-size: 2.5rem;
+  font-size: clamp(1.8rem, 4vw, 2.5rem);
   color: white;
   font-weight: bold;
   margin-bottom: 0.2rem;
@@ -38,12 +38,10 @@
   box-shadow: 0 0 20px rgba(255, 255, 255, 0.2);
   border: 4px solid #374151;
   border-radius: 10px;
-  /* overflow: hidden; */
-   max-width: 900px; /* Add this */
   position: relative;
+  max-width: 900px;
 }
 
-/* Square Styles */
 .square {
   display: flex;
   align-items: center;
@@ -71,12 +69,8 @@
 }
 
 .possible-move {
-  outline: 3px dashed limegreen;
-  outline-offset: -2px;
-}
-
-.square.possible-move {
   outline: 3px solid #2e7d32;
+  outline-offset: -2px;
 }
 
 /* Piece Icons */
@@ -84,6 +78,27 @@
   width: 60px;
   height: 60px;
   object-fit: contain;
+}
+
+.square img {
+  opacity: 0;
+  transform: scale(0.95);
+  animation: pieceFadeIn 0.3s ease forwards;
+}
+
+@keyframes pieceFadeIn {
+  0% {
+    opacity: 0;
+    transform: scale(0.85);
+  }
+  100% {
+    opacity: 1;
+    transform: scale(1);
+  }
+}
+
+.square img:active {
+  transform: scale(0.9);
 }
 
 /* Result Banner */
@@ -97,6 +112,7 @@
   border-radius: 8px;
   text-align: center;
   display: inline-block;
+  z-index: 15;
 }
 
 /* Refresh Icon */
@@ -113,7 +129,7 @@
   transform: rotate(90deg);
 }
 
-/* For overlays like THINKING... */
+/* Relative Container */
 .relative {
   display: flex;
   justify-content: center;
@@ -147,97 +163,25 @@
 
 .promotion-option:hover {
   transform: scale(1.1);
-}
-
-/* Optional light-themed modal */
-.dialog-content-light {
-  background-color: #ffffff;
-  color: #000000;
-  border-radius: 0.75rem;
-  padding: 1.5rem;
-  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
-}
-
-/* Mobile Responsive Adjustments */
-@media (max-width: 500px) {
-  .board {
-    grid-template-columns: repeat(8, 45px);
-    grid-template-rows: repeat(8, 45px);
-  }
-
-  .piece {
-    width: 40px;
-    height: 40px;
-  }
-
-  .title {
-    font-size: 1.8rem;
-    margin-bottom: 1rem;
-  }
-
-  .result-banner {
-    font-size: 1rem;
-    padding: 0.75rem 1.5rem;
-  }
-
-  .new-game-icon {
-    width: 1.2rem;
-    height: 1.2rem;
-  }
-
-  .promotion-option {
-    width: 45px;
-    height: 45px;
-  }
-}
-.promotion-option {
-  width: 60px;
-  height: 60px;
-  cursor: pointer;
-  border: 2px solid white;
-  border-radius: 0.25rem;
-  transition: transform 0.2s;
-}
-
-.promotion-option:hover {
-  transform: scale(1.1);
-  border-color: #facc15; /* Tailwind yellow-400 */
+  border-color: #facc15;
   box-shadow: 0 0 10px #facc15;
 }
-/* Add this at the bottom of ChessGame.css */
 
+/* Chess Background */
 .chess-background {
   position: fixed;
   top: 0;
   left: 0;
   width: 100vw;
   height: 100vh;
-  background-image: url("../../assets/chess3.jpg"); /* Adjust path if needed */
+  background-image: url("../../assets/chess3.jpg");
   background-size: cover;
   background-position: center;
   filter: blur(10px) brightness(0.5);
   z-index: -1;
 }
-.square img {
-  opacity: 0;
-  transform: scale(0.95);
-  animation: pieceFadeIn 0.3s ease forwards;
-}
 
-@keyframes pieceFadeIn {
-  0% {
-    opacity: 0;
-    transform: scale(0.85);
-  }
-  100% {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
-
-.square img:active {
-  transform: scale(0.9);
-}
+/* Board Container and Captured Pieces */
 .board-container {
   display: flex;
   justify-content: center;
@@ -245,7 +189,7 @@
   gap: 1rem;
   margin-top: 1rem;
   position: relative;
-  z-index: 1;
+  z-index: 10;
 }
 
 .captured-column {
@@ -274,3 +218,94 @@
   filter: grayscale(0%);
 }
 
+/* Mobile Responsive Adjustments */
+@media (max-width: 640px) {
+  .board-container {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .captured-column {
+    flex-direction: row;
+    justify-content: center;
+    flex-wrap: wrap;
+    max-height: none;
+    max-width: 100%;
+    background-color: transparent;
+    box-shadow: none;
+    gap: 0.4rem;
+    padding: 0.25rem;
+  }
+
+  .captured-icon {
+    width: 30px;
+    height: 30px;
+  }
+
+  .board {
+    grid-template-columns: repeat(8, 45px);
+    grid-template-rows: repeat(8, 45px);
+  }
+
+  .piece {
+    width: 40px;
+    height: 40px;
+  }
+
+  .promotion-option {
+    width: 45px;
+    height: 45px;
+  }
+
+  .title {
+    font-size: 1.8rem;
+    margin-bottom: 1rem;
+  }
+
+  .result-banner {
+    font-size: 1rem;
+    padding: 0.75rem 1.5rem;
+  }
+
+  .new-game-icon {
+    width: 1.2rem;
+    height: 1.2rem;
+  }
+}
+
+@media (max-width: 400px) {
+  .board {
+    grid-template-columns: repeat(8, 38px);
+    grid-template-rows: repeat(8, 38px);
+  }
+
+  .piece {
+    width: 35px;
+    height: 35px;
+  }
+
+  .promotion-option {
+    width: 40px;
+    height: 40px;
+  }
+
+  .title {
+    font-size: 1.6rem;
+  }
+
+  .result-banner {
+    font-size: 0.9rem;
+    padding: 0.6rem 1rem;
+  }
+
+  .new-game-icon {
+    width: 1rem;
+    height: 1rem;
+  }
+
+  .relative .text-5xl {
+    font-size: 1.5rem;
+    text-align: center;
+    padding: 0 0.5rem;
+  }
+}
