+++ src/pages/startgame/StartGame.styles.ts
@@ -13,9 +13,10 @@ export const overlayStyle = cn(
 );
 
 export const headingStyle = cn(
-  "text-4xl md:text-5xl font-bold text-white drop-shadow-md"
+  "text-2xl sm:text-3xl md:text-5xl font-bold text-white drop-shadow-md text-center px-4"
 );
 
 export const iconButtonStyle = cn(
-  "rounded-full w-16 h-16 bg-yellow-400 hover:bg-yellow-500 text-black flex items-center justify-center shadow-lg transition transform hover:scale-105"
+  "rounded-full w-12 h-12 sm:w-14 sm:h-14 md:w-16 md:h-16 bg-yellow-400 hover:bg-yellow-500 text-black flex items-center justify-center shadow-lg transition transform hover:scale-105"
 );
+
